CMakeLists: fix compilation with Ninja build system
On using Ninja build system, the package currently don't configure
correctly with the error:

ninja: error: 'check/*', needed by 'CMakeFiles/radiotap_check', missing and no known rule to make it

This is caused by the usage of

DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/check/*

that both CMake and Ninja doesn't expand to the list of all files in the
check directory.

To better address this, use the CMake GLOB to first enumerate all the
files in the check directory and then call DEPENDS on a variable
containing the list of files.

This restore correct functionality and better adapt to CMake syntax.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>